Renamed apartment and tenant files to be singular. Still and issue with Json and now and issue with 'require_relative'
