+++ db/models/apartment.rb
